Update the generation of the PerfCounter event
The counter carries not a running sum any more, but the number of samples
since last event. It records the OS thread so that we can try and assign
a cap to a counter. In all events the type of OS thread IDs is changed
to Word64, after Duncan discovered that some OSes need that.