Retain the sync-by-syscall code variant on a branch
This version of the code uses the linux-perf library to parse
the perf data file and an artificially inserted nanosleep syscall
to synchronize perf events and the standard eventlog events.
This approach has its benefits and drawbacks compared to the usage
of perf-script for both tasks. Which of the approaches proves better depends
on the evolution of the perf toolset (we are waiting for a patchset
that makes perf clock trustworthy and externally meaningful) and on the
future needs of Haskell users of perf events. For now, we retain
this version on a branch.